fixed deletion and fixed props

+++ backend/src/bullmq/workers/delete.worker.ts
@@ -21,6 +21,7 @@ import {v2 as cloudinary} from "cloudinary"
 import fileModel from "../../models/file.model.js";
 import { QdrantClient } from "@qdrant/js-client-rest";
 import messageModel from "../../models/messages.model.js";
+import urlModel from "../../models/url.model.js";
 
 cloudinary.config({
     cloud_name:process.env.CLOUDINARY_CLOUD_NAME,
@@ -33,20 +34,15 @@ await DB();
 const worker = new Worker('delete-file-queue',async (job:Job)=>{
     console.log("started Worker");
     try {
-        const {id , userId , qdrantCollection , publicId} = job.data;
+        const {id , userId , qdrantCollection , publicId , sourceType} = job.data;
 
-        if(!id || !userId || !publicId){
+        if(!id || !userId || !sourceType || !qdrantCollection){
             throw new Error("FileId or userId not found")
         }
         console.log("FileId:",id);
         console.log("UserId:",userId);
+        console.log("QdrantCollection:",qdrantCollection);
         // console.log("PublicId:",publicId);
-        
-        const qdrantCollectionDeleted = await client.deleteCollection(qdrantCollection)
-        if(!qdrantCollectionDeleted){
-            throw new Error("Error deleted qdrant collection")
-        }
-        console.log('Qdrant collection deleted');
         
         if(publicId){
             const cloudinaryDeletion = await cloudinary.uploader.destroy(publicId,{resource_type:'raw',invalidate:true})
@@ -54,7 +50,15 @@ const worker = new Worker('delete-file-queue',async (job:Job)=>{
                 throw new Error("Error deletion cloudinary file")
             }
             console.log('Cloudinary file deleted',cloudinaryDeletion);
+        }else{
+            console.log("No public ID ----------Skipping")
         }
+        
+        const qdrantCollectionDeleted = await client.deleteCollection(qdrantCollection)
+        if(!qdrantCollectionDeleted){
+            throw new Error("Error deleted qdrant collection")
+        }
+        console.log('Qdrant collection deleted');
         
         const messagesDeleted = await messageModel.deleteMany({id,userId});
         if(!messagesDeleted){
@@ -62,11 +66,19 @@ const worker = new Worker('delete-file-queue',async (job:Job)=>{
         }
         console.log('User messages deleted');
         
-        const mongoDBFile = await fileModel.deleteOne({_id:id,userId})
-        if(!mongoDBFile){
-            throw new Error("Error deleting MongoDB file")
+        if(sourceType==="file"){
+            const mongoDBFile = await fileModel.deleteOne({_id:id,userId})
+            if(!mongoDBFile){
+                throw new Error("Error deleting MongoDB file")
+            }
+            console.log('MonogDB file deleted');
+        }else{
+            const mongoDBFile = await urlModel.deleteOne({_id:id,userId})
+            if(!mongoDBFile){
+                throw new Error("Error deleting MongoDB file")
+            }
+            console.log('MonogDB Url deleted');
         }
-        console.log('MonogDB file deleted');
 
     } catch (error) {
         console.error("Worker job failed:", error);
